Refactor configuration validation by moving it from a separate helper file to the main connector module. Enhance validation logic to include checks for optional parameters such as search engine, response format, country, and search zone, ensuring all values meet expected types and constraints. Update README to reflect changes in the validation function's location.
